+++ nassh/js/nassh_main.js
@@ -42,6 +42,15 @@ const openNewWindow = function(url) {
  * so we do it like this instead.
  */
 globalThis.addEventListener('DOMContentLoaded', async (event) => {
+  // If we're being opened by a link from another page, clear the opener setting
+  // so we can't reach back into them.  They should have used noopener, but help
+  // cover if they don't.
+  if (globalThis.opener !== null) {
+    globalThis.opener = null;
+    globalThis.location.reload();
+    return;
+  }
+
   // Check if site's storage has been marked as persistent.
   if (globalThis.navigator?.storage?.persist &&
       globalThis.navigator?.storage?.persisted) {